# docs/swagger.yaml
definitions:
...
  errInfos.ErrCode:
    enum:
    - 10001
    - 10002
    - 10003
    - 10004
    - 10005
    - 10006
    - 10007
    - 10008
    - 10009
    - 20001
    - 20002
    - 30001
    - 30002
    - 30003
    - 30004
    - 30005
    - 40001
    - 40002
    - 40003
    - 40004
    - 40005
    - 40006
    - 40007
    - 40008
    - 40009
    - 50001
    - 50002
    - 50003
    - 50004
    - 50005
    - 50006
    - 50007
    - 50008
    - 60001
    - 60002
    - 60003
    - 60004
    - 60005
    - 70001
    - 70002
    - 70003
    - 70004
    - 80001
    - 90001
    - 90002
    - 90003
    - 90004
    - 90005
    - 100001
    - 100002
    - 100003
    - 100004
    - 110001
    - 110002
    - 110003
    - 120001
    - 120002
    - 120003
    - 120004
    - 120005
    - 120006
    - 120007
    - 120008
    - 120009
    - 120010
    - 120011
    - 120012
    - 50009
    - 50010
    - 50011
    - 50012
    - 50013
    - 50014
    - 50015
    - 50016
    - 50017
    - 50018
    - 60006
    - 60007
    - 60008
    - 60009
    - 60010
    - 60011
    - 60012
    - 60013
    - 60014
    - 60015
    - 60016
    - 60017
    - 60018
    type: integer
    x-enum-comments:
      ERR_CHECK_VIOLATION: CHECK 約束違反
      ERR_CONCURRENT_MODIFIED: 並發修改衝突
      ERR_CONSTRAINT_VIOLATION: 違反資料庫約束
      ERR_DEADLOCK_DETECTED: 偵測到資料庫死鎖
      ERR_EXCLUSIVE_LOCK_FAILED: 排他鎖獲取失敗
      ERR_FOREIGN_KEY_VIOLATION: 外鍵約束違反
      ERR_LOCK_WAIT_TIMEOUT: 鎖等待超時
      ERR_PARTIAL_COMPLETION: 交易部分完成（部分操作成功）
      ERR_RESOURCE_LOCKED: 資源被鎖定（如另一筆交易正在修改）
      ERR_ROLLBACK_FAILED: 交易回滾失敗
      ERR_SERIALIZATION_FAILURE: 序列化失敗（並發衝突）
      ERR_SHARE_LOCK_FAILED: 共享鎖獲取失敗
      ERR_TX_FAILED: 交易執行失敗
      ERR_TX_TIMEOUT: 交易執行超時
      ERR_UNIQUE_VIOLATION: 唯一約束違反
      EXCEPTION_ALREADY_PROCESSED: 例外已被處理
      EXCEPTION_CANCEL_DEADLINE_PASSED: 停課截止日已過
      EXCEPTION_DEADLINE_EXCEEDED: 超過例外申請截止日
      EXCEPTION_REPLACE_TEACHER_INVALID: 代課老師無效
      EXCEPTION_RESCHEDULE_CONFLICT: 調課時間與現有排程衝突
      EXCEPTION_RESCHEDULE_NO_NEW_TIME: 調課必須提供新時間
      EXCEPTION_SELF_REVIEW_FORBIDDEN: 不能審核自己提交的申請
      RECURRENCE_BATCH_LIMIT_EXCEEDED: 批量操作超過限制
      RECURRENCE_DELETE_CONFIRM: 刪除操作需要確認
      RECURRENCE_EDIT_DATE_REQUIRED: 編輯日期為必填
      RECURRENCE_EDIT_MODE_INVALID: 無效的編輯模式
      RECURRENCE_FUTURE_WITH_EDIT_DATE: FUTURE 模式必須指定編輯日期
      RECURRENCE_NO_AFFECTED_SESSIONS: 沒有受影響的場次
      SCHED_COURSE_NOT_FOUND: 課程模板不存在
      SCHED_DURATION_EXCEEDS_LIMIT: 課程時長超過限制
      SCHED_END_BEFORE_START: 結束日期早於開始日期
      SCHED_INVALID_DATE_FORMAT: 無效的日期格式
      SCHED_INVALID_DURATION: 無效的課程時長
      SCHED_INVALID_WEEKDAY: 無效的星期幾
      SCHED_OFFERING_NOT_FOUND: 班別不存在
      SCHED_ROOM_REQUIRED: 必須指定教室
      SCHED_START_AFTER_END: 開始時間晚於結束時間
      SCHED_TEACHER_REQUIRED: 必須指定老師
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - ""
    - 資源被鎖定（如另一筆交易正在修改）
    - 並發修改衝突
    - 交易執行失敗
    - 交易執行超時
    - 偵測到資料庫死鎖
    - 交易部分完成（部分操作成功）
    - 交易回滾失敗
    - 違反資料庫約束
    - 外鍵約束違反
    - 唯一約束違反
    - CHECK 約束違反
    - 序列化失敗（並發衝突）
    - 鎖等待超時
    - 共享鎖獲取失敗
    - 排他鎖獲取失敗
    - 必須指定老師
    - 必須指定教室
    - 班別不存在
    - 課程模板不存在
    - 無效的星期幾
    - 無效的課程時長
    - 開始時間晚於結束時間
    - 無效的日期格式
    - 結束日期早於開始日期
    - 課程時長超過限制
    - 超過例外申請截止日
    - 不能審核自己提交的申請
    - 例外已被處理
    - 調課時間與現有排程衝突
    - 代課老師無效
    - 停課截止日已過
    - 調課必須提供新時間
    - 無效的編輯模式
    - 沒有受影響的場次
    - FUTURE 模式必須指定編輯日期
    - 編輯日期為必填
    - 刪除操作需要確認
    - 批量操作超過限制
    x-enum-varnames:
    - SYSTEM_ERROR
    - PARAMS_VALIDATE_ERROR
    - JSON_ENCODE_ERROR
    - JSON_DECODE_ERROR
    - JSON_PROCESS_ERROR
    - FORMAT_RESOURCE_ERROR
    - REQUEST_TIMEOUT
    - NOT_IMPLEMENTED
    - RATE_LIMIT_EXCEEDED
    - SQL_ERROR
    - TX_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - TOKEN_EXPIRED
    - INVALID_TOKEN
    - INVALID_INVITE
    - NOT_FOUND
    - DUPLICATE
    - TAG_INVALID
    - LIMIT_EXCEEDED
    - RESOURCE_IN_USE
    - COURSE_IN_USE
    - OFFERING_HAS_RULES
    - ROOM_IN_USE
    - INVALID_STATUS
    - SCHED_OVERLAP
    - SCHED_BUFFER
    - SCHED_PAST
    - SCHED_LOCKED
    - SCHED_CLOSED
    - SCHED_INVALID_RANGE
    - SCHED_RULE_CONFLICT
    - SCHED_EXCEPTION_EXISTS
    - EXCEPTION_NOT_FOUND
    - EXCEPTION_INVALID_ACTION
    - EXCEPTION_REVIEWED
    - EXCEPTION_REVOKED
    - EXCEPTION_REJECT_SELF
    - FILE_TOO_LARGE
    - FILE_TYPE_INVALID
    - UPLOAD_FAILED
    - CERTIFICATE_NOT_FOUND
    - TALENT_NOT_OPEN
    - LINE_ALREADY_BOUND
    - LINE_NOT_BOUND
    - LINE_BINDING_CODE_INVALID
    - LINE_BINDING_EXPIRED
    - LINE_NOTIFY_FAILED
    - ADMIN_NOT_FOUND
    - ADMIN_EMAIL_EXISTS
    - PASSWORD_NOT_MATCH
    - ADMIN_CANNOT_DISABLE_SELF
    - ERR_RESOURCE_LOCKED
    - ERR_CONCURRENT_MODIFIED
    - ERR_TX_FAILED
    - ERR_TX_TIMEOUT
    - ERR_DEADLOCK_DETECTED
    - ERR_PARTIAL_COMPLETION
    - ERR_ROLLBACK_FAILED
    - ERR_CONSTRAINT_VIOLATION
    - ERR_FOREIGN_KEY_VIOLATION
    - ERR_UNIQUE_VIOLATION
    - ERR_CHECK_VIOLATION
    - ERR_SERIALIZATION_FAILURE
    - ERR_LOCK_WAIT_TIMEOUT
    - ERR_SHARE_LOCK_FAILED
    - ERR_EXCLUSIVE_LOCK_FAILED
    - SCHED_TEACHER_REQUIRED
    - SCHED_ROOM_REQUIRED
    - SCHED_OFFERING_NOT_FOUND
    - SCHED_COURSE_NOT_FOUND
    - SCHED_INVALID_WEEKDAY
    - SCHED_INVALID_DURATION
    - SCHED_START_AFTER_END
    - SCHED_INVALID_DATE_FORMAT
    - SCHED_END_BEFORE_START
    - SCHED_DURATION_EXCEEDS_LIMIT
    - EXCEPTION_DEADLINE_EXCEEDED
    - EXCEPTION_SELF_REVIEW_FORBIDDEN
    - EXCEPTION_ALREADY_PROCESSED
    - EXCEPTION_RESCHEDULE_CONFLICT
    - EXCEPTION_REPLACE_TEACHER_INVALID
    - EXCEPTION_CANCEL_DEADLINE_PASSED
    - EXCEPTION_RESCHEDULE_NO_NEW_TIME
    - RECURRENCE_EDIT_MODE_INVALID
    - RECURRENCE_NO_AFFECTED_SESSIONS
    - RECURRENCE_FUTURE_WITH_EDIT_DATE
    - RECURRENCE_EDIT_DATE_REQUIRED
    - RECURRENCE_DELETE_CONFIRM
    - RECURRENCE_BATCH_LIMIT_EXCEEDED
  global.ApiResponse:
    properties:
      code:
        $ref: '#/definitions/errInfos.ErrCode'
      datas: {}
      message:
        type: string
      warnings:
        items:
          $ref: '#/definitions/global.Warning'
        type: array
    type: object
...
  global.Warning:
    properties:
      details:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
...
  resources.CourseResponse:
    properties:
      center_id:
        type: integer
      color_hex:
        type: string
      created_at:
        type: string
      default_duration:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      room_buffer_min:
        type: integer
      teacher_buffer_min:
        type: integer
    type: object
...
  resources.RoomResponse:
    properties:
      capacity:
        type: integer
      center_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  services.CreateCourseRequest:
    properties:
      color_hex:
        type: string
      duration:
        type: integer
      name:
        type: string
      room_buffer_min:
        minimum: 0
        type: integer
      teacher_buffer_min:
        minimum: 0
        type: integer
    required:
    - color_hex
    - duration
    - name
    type: object
...
  services.CreateRoomRequest:
    properties:
      capacity:
        type: integer
      name:
        type: string
    required:
    - capacity
    - name
    type: object
...
  services.ToggleActiveRequest:
    properties:
      is_active:
        type: boolean
    type: object
  services.UpdateCourseRequest:
    properties:
      color_hex:
        type: string
      duration:
        type: integer
      name:
        type: string
      room_buffer_min:
        minimum: 0
        type: integer
      teacher_buffer_min:
        minimum: 0
        type: integer
    required:
    - color_hex
    - duration
    - name
    type: object
  services.UpdateRoomRequest:
    properties:
      capacity:
        type: integer
      name:
        type: string
    required:
    - capacity
    - name
    type: object
...
info:
  contact: {}
  description: API 維護文件
  title: timeLedger 阿卡莉 模板框架
  version: "1.0"
paths:
...
  /api/v1/admin/courses:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resources.CourseResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得課程列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: 課程資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/resources.CourseResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新增課程
      tags:
      - Admin
  /api/v1/admin/courses/{course_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.ApiResponse'
      security:
      - BearerAuth: []
      summary: 刪除課程
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: 課程資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/resources.CourseResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新課程
      tags:
      - Admin
  /api/v1/admin/courses/{course_id}/toggle-active:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: 啟用狀態
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.ToggleActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.ApiResponse'
      security:
      - BearerAuth: []
      summary: 切換課程啟用狀態
      tags:
      - Admin
  /api/v1/admin/courses/active:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resources.CourseResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得已啟用的課程列表
      tags:
      - Admin
...
  /api/v1/admin/rooms:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resources.RoomResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得教室列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: 教室資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/resources.RoomResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新增教室
      tags:
      - Admin
  /api/v1/admin/rooms/{room_id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      - description: 教室資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/resources.RoomResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新教室
      tags:
      - Admin
  /api/v1/admin/rooms/{room_id}/toggle-active:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      - description: 啟用狀態
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.ToggleActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.ApiResponse'
      security:
      - BearerAuth: []
      summary: 切換教室啟用狀態
      tags:
      - Admin
  /api/v1/admin/rooms/active:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resources.RoomResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得已啟用的教室列表
      tags:
      - Admin
...
schemes:
- http
swagger: "2.0"